% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tools.MLE}
\alias{tools.MLE}
\title{getting Maximum Likelihood Estimator}
\usage{
tools.MLE(data, dist = c("bin", "exp", "norm", "pois"))
}
\arguments{
\item{data}{the dataset to be used}

\item{dist}{type of distribution of the dataset}
}
\value{
Maximum Likelihood Estimator
}
\description{
getting Maximum Likelihood Estimator
}
