% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{toolsMLE.l}
\alias{toolsMLE.l}
\title{getting likelihood value}
\usage{
toolsMLE.l(data, dist = c("bin", "exp", "norm", "pois"))
}
\arguments{
\item{data}{the dataset to be used}

\item{dist}{type of distribution of the dataset}
}
\value{
likelihood value
}
\description{
getting likelihood value
}
