% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{toolsMLE.ll}
\alias{toolsMLE.ll}
\title{getting log-likelihood value}
\usage{
toolsMLE.ll(data, dist = c("bin", "exp", "norm", "pois"))
}
\arguments{
\item{data}{the dataset to be used}

\item{dist}{type of distribution of the dataset}
}
\value{
log-likelihood value
}
\description{
getting log-likelihood value
}
