% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{toolsMLE.lrt}
\alias{toolsMLE.lrt}
\title{conducting likelihood ratio test}
\usage{
toolsMLE.lrt(data, dist = c("bin", "exp", "norm", "pois"))
}
\arguments{
\item{data}{the dataset to be used}

\item{dist}{type of distribution of the dataset}
}
\value{
likelihood ratio test statistic and p-value
}
\description{
conducting likelihood ratio test
}
